+++ traknor/presentation/reports/tests.py
@@ -13,7 +13,7 @@ def _create_user():
         password="pass",
         first_name="Test",
         last_name="User",
-        role="TECH",
+        role="technician",
     )